Drop w3m and use html2text
After researching for a while, I found there is no a good way to produce
same output as what w3m generates. Instead, html2text should be a
helpful to convert HTML to markdown. This would be a good way for the
future changes that to write case text in markdown directly, which is
not determined yet though.

Fix #25
