--- tcms/core/utils/html.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-from subprocess import Popen, PIPE
-
-
-def html2text(value):
-    """
-    Pipes given HTML string into the text browser W3M, which renders it.
-    Rendered text is grabbed from STDOUT and returned.
-    """
-    try:
-        cmd = "w3m -dump -T text/html -O ascii"
-        proc = Popen(cmd, shell=True, stdin=PIPE, stdout=PIPE)
-        return proc.communicate(str(value))[0]
-    except OSError:
-        # something bad happened, so just return the input
-        return value
