+++ tcms/testcases/tests/test_views.py
@@ -17,8 +17,8 @@ from django.test import RequestFactory
 
 from uuslug import slugify
 
-from .fields import MultipleEmailField
-from .forms import CaseTagForm
+from ..fields import MultipleEmailField
+from ..forms import CaseTagForm
 from tcms.management.models import TestTag
 from tcms.testcases.models import TestCase
 from tcms.testcases.models import TestCaseTag
@@ -27,14 +27,11 @@ from tcms.testcases.models import TestCaseComponent
 from tcms.testcases.models import TestCasePlan
 from tcms.testcases.views import ajax_response
 from tcms.tests.factories import ComponentFactory
-from tcms.tests.factories import TestBuildFactory
 from tcms.tests.factories import TestCaseCategoryFactory
 from tcms.tests.factories import TestCaseComponentFactory
 from tcms.tests.factories import TestCaseFactory
-from tcms.tests.factories import TestCaseRunFactory
 from tcms.tests.factories import TestCaseTagFactory
 from tcms.tests.factories import TestPlanFactory
-from tcms.tests.factories import TestRunFactory
 from tcms.tests.factories import TestTagFactory
 from tcms.tests import BasePlanCase
 from tcms.tests import remove_perm_from_user
@@ -87,146 +84,6 @@ class TestMultipleEmailField(unittest.TestCase):
         self.field.required = False
         data = self.field.clean(value)
         self.assertEqual(data, [])
-
-
-class TestCaseRemoveBug(BasePlanCase):
-    """Test TestCase.remove_bug"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super(TestCaseRemoveBug, cls).setUpTestData()
-        cls.build = TestBuildFactory(product=cls.product)
-        cls.test_run = TestRunFactory(product_version=cls.version, plan=cls.plan,
-                                      manager=cls.tester, default_tester=cls.tester)
-        cls.case_run = TestCaseRunFactory(assignee=cls.tester, tested_by=cls.tester,
-                                          case=cls.case, run=cls.test_run, build=cls.build)
-        cls.bug_system = TestCaseBugSystem.objects.get(name='Bugzilla')
-
-    def setUp(self):
-        self.bug_id_1 = '12345678'
-        self.case.add_bug(self.bug_id_1, self.bug_system.pk,
-                          summary='error when add a bug to a case')
-        self.bug_id_2 = '10000'
-        self.case.add_bug(self.bug_id_2, self.bug_system.pk, case_run=self.case_run)
-
-    def tearDown(self):
-        self.case.case_bug.all().delete()
-
-    def test_remove_case_bug(self):
-        self.case.remove_bug(self.bug_id_1)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_1).exists()
-        self.assertFalse(bug_found)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_2).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_2))
-
-    def test_case_bug_not_removed_by_passing_case_run(self):
-        self.case.remove_bug(self.bug_id_1, run_id=self.case_run.pk)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_1).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_1))
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_2).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_2))
-
-    def test_remove_case_run_bug(self):
-        self.case.remove_bug(self.bug_id_2, run_id=self.case_run.pk)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_2).exists()
-        self.assertFalse(bug_found)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_1).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_1))
-
-    def test_case_run_bug_not_removed_by_missing_case_run(self):
-        self.case.remove_bug(self.bug_id_2)
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_1).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_1))
-
-        bug_found = self.case.case_bug.filter(bug_id=self.bug_id_2).exists()
-        self.assertTrue(bug_found,
-                        'Bug {0} does not exist. It should not be deleted.'.format(self.bug_id_2))
-
-
-class TestCaseRemoveComponent(BasePlanCase):
-    """Test TestCase.remove_component"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super(TestCaseRemoveComponent, cls).setUpTestData()
-
-        cls.component_1 = ComponentFactory(name='Application',
-                                           product=cls.product,
-                                           initial_owner=cls.tester,
-                                           initial_qa_contact=cls.tester)
-        cls.component_2 = ComponentFactory(name='Database',
-                                           product=cls.product,
-                                           initial_owner=cls.tester,
-                                           initial_qa_contact=cls.tester)
-
-        cls.cc_rel_1 = TestCaseComponentFactory(case=cls.case,
-                                                component=cls.component_1)
-        cls.cc_rel_2 = TestCaseComponentFactory(case=cls.case,
-                                                component=cls.component_2)
-
-    def test_remove_a_component(self):
-        self.case.remove_component(self.component_1)
-
-        found = self.case.component.filter(pk=self.component_1.pk).exists()
-        self.assertFalse(
-            found,
-            'Component {0} exists. But, it should be removed.'.format(
-                self.component_1.pk))
-        found = self.case.component.filter(pk=self.component_2.pk).exists()
-        self.assertTrue(
-            found,
-            'Component {0} does not exist. It should not be removed.'.format(
-                self.component_2.pk))
-
-
-class TestCaseRemovePlan(BasePlanCase):
-    """Test TestCase.remove_plan"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super(TestCaseRemovePlan, cls).setUpTestData()
-
-        cls.new_case = TestCaseFactory(author=cls.tester, default_tester=None, reviewer=cls.tester,
-                                       plan=[cls.plan])
-
-    def test_remove_plan(self):
-        self.new_case.remove_plan(self.plan)
-
-        found = self.plan.case.filter(pk=self.new_case.pk).exists()
-        self.assertFalse(
-            found, 'Case {0} should has no relationship to plan {0} now.'.format(self.new_case.pk,
-                                                                                 self.plan.pk))
-
-
-class TestCaseRemoveTag(BasePlanCase):
-    """Test TestCase.remove_tag"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super(TestCaseRemoveTag, cls).setUpTestData()
-
-        cls.tag_rhel = TestTagFactory(name='rhel')
-        cls.tag_fedora = TestTagFactory(name='fedora')
-        TestCaseTagFactory(case=cls.case, tag=cls.tag_rhel, user=cls.tester.pk)
-        TestCaseTagFactory(case=cls.case, tag=cls.tag_fedora, user=cls.tester.pk)
-
-    def test_remove_tag(self):
-        self.case.remove_tag(self.tag_rhel)
-
-        tag_pks = list(self.case.tag.all().values_list('pk', flat=True))
-        self.assertEqual([self.tag_fedora.pk], tag_pks)
 
 
 class CaseTagFormTest(test.TestCase):
